examples/petpassport/petpassport/requirements.txt
google-adk==1.28.0
python-dotenv
google-genai
pillow
reportlab
